requirements.txt
openai>=1.0
python-dotenv
gitpython
requests
tqdm